feat: define protocol contract and update IPC handling
