+++ gems/vajra/ext/vajra/ipc/frame_contract.hpp
@@ -6,26 +6,6 @@
 #ifndef VAJRA_IPC_FRAME_CONTRACT_HPP
 #define VAJRA_IPC_FRAME_CONTRACT_HPP
 
-#include "channel_kind.hpp"
-#include "frame_family.hpp"
-#include "protocol_version.hpp"
-
-#include <array>
-#include <cstdint>
-#include <optional>
-
-namespace Vajra
-{
-  namespace ipc
-  {
-    bool known_frame_family(FrameFamily family);
-    std::optional<ProtocolVersion> first_supported_protocol_version(FrameFamily family);
-    ChannelKind owning_channel(FrameFamily family);
-    bool valid_on_channel(FrameFamily family, ChannelKind channel);
-    bool reserved_family(FrameFamily family);
-    std::uint16_t wire_id(FrameFamily family);
-    std::optional<FrameFamily> decode_frame_family(std::uint16_t wire_id);
-  }
-}
+#include "protocol_contract.hpp"
 
 #endif
